requirements.txt
asgiref
Brotli
dj-database-url
Django
gunicorn
psycopg2-binary
sqlparse
tzdata
whitenoise
pytz
djangorestframework
